+++ packages/tenderly-core/src/internal/core/services/TenderlyService.ts
@@ -1,5 +1,4 @@
 import { TENDERLY_DASHBOARD_BASE_URL, CHAIN_ID_NETWORK_NAME_MAP } from "../../../common/constants";
-import { logApiError } from "../common/logger";
 import {
   API_VERIFICATION_REQUEST_ERR_MSG,
   BYTECODE_MISMATCH_ERR_MSG,
@@ -19,7 +18,14 @@ import {
   TenderlyContractUploadRequest,
   TenderlyForkContractUploadRequest,
 } from "../types";
+import { logger } from "../../../utils/logger";
 import { TenderlyApiService } from "./TenderlyApiService";
+import {
+  convertToLogCompliantApiError,
+  convertToLogCompliantNetworks,
+  convertToLogCompliantProjects,
+  convertToLogCompliantVerificationResponse
+} from "../../../utils/log-compliance";
 
 export class TenderlyService {
   private pluginName: string;
@@ -29,6 +35,8 @@ export class TenderlyService {
   }
 
   public async getNetworks(): Promise<TenderlyNetwork[]> {
+    logger.debug("Obtaining public networks.");
+
     let tenderlyApi = TenderlyApiService.configureAnonymousInstance();
     if (TenderlyApiService.isAuthenticated()) {
       tenderlyApi = TenderlyApiService.configureInstance();
@@ -36,15 +44,25 @@ export class TenderlyService {
 
     try {
       const res = await tenderlyApi.get("/api/v1/public-networks");
+      if (res.data === undefined || res.data === null) {
+        logger.error("There was an error while obtaining public networks from Tenderly. Obtained response is invalid.");
+        return [];
+      }
+      const logCompliantNetworks = convertToLogCompliantNetworks(res.data);
+      logger.silly("Obtained public networks:", logCompliantNetworks);
+
       return res.data;
     } catch (err) {
-      logApiError(err);
+      const logCompliantApiErr = convertToLogCompliantApiError(err);
+      logger.error(logCompliantApiErr);
       console.log(`Error in ${this.pluginName}: ${NETWORK_FETCH_FAILED_ERR_MSG}`);
     }
     return [];
   }
 
   public async getLatestBlockNumber(networkId: string): Promise<string | null> {
+    logger.debug("Getting latest block number.");
+
     let tenderlyApi = TenderlyApiService.configureAnonymousInstance();
     if (TenderlyApiService.isAuthenticated()) {
       tenderlyApi = TenderlyApiService.configureInstance();
@@ -52,15 +70,26 @@ export class TenderlyService {
 
     try {
       const res = await tenderlyApi.get(`/api/v1/network/${networkId}/block-number`);
+      if (res.data === undefined || res.data === null) {
+        logger.error(
+          "There was an error while obtaining latest block number from Tenderly. Obtained response is invalid."
+        );
+        return null;
+      }
+      logger.trace(`Api successfully returned: ${res.data.block_number}`);
+
       return res.data.block_number;
     } catch (err) {
-      logApiError(err);
-      console.log(`Error in ${this.pluginName}: ${LATEST_BLOCK_NUMBER_FETCH_FAILED_ERR_MSG}`);
+      const logCompliantApiErr = convertToLogCompliantApiError(err);
+      logger.error(logCompliantApiErr);
+      logger.error(`Error in ${this.pluginName}: ${LATEST_BLOCK_NUMBER_FETCH_FAILED_ERR_MSG}`);
     }
     return null;
   }
 
   public async verifyContracts(request: TenderlyContractUploadRequest): Promise<void> {
+    logger.debug("Verifying contracts publicly.");
+
     let tenderlyApi = TenderlyApiService.configureAnonymousInstance();
     if (TenderlyApiService.isAuthenticated()) {
       tenderlyApi = TenderlyApiService.configureInstance();
@@ -68,15 +97,28 @@ export class TenderlyService {
 
     try {
       if (request.contracts.length === 0) {
-        console.log(NO_VERIFIABLE_CONTRACTS_ERR_MSG);
+        logger.error(NO_VERIFIABLE_CONTRACTS_ERR_MSG);
         return;
       }
 
       const res = await tenderlyApi.post("/api/v1/public/verify-contracts", { ...request });
+      if (res.data === undefined || res.data === null) {
+        logger.error(
+          "There was an error while publicly verifying contracts on Tenderly. Obtained response is invalid."
+        );
+        return;
+      }
+      const logCompliantVerificationResponse = convertToLogCompliantVerificationResponse(res.data);
+      logger.trace("Verification response:", logCompliantVerificationResponse);
 
       const responseData: ContractResponse = res.data;
       if (responseData.bytecode_mismatch_errors !== null) {
-        console.log(`Error in ${this.pluginName}: ${BYTECODE_MISMATCH_ERR_MSG}`);
+        logger.error(`Error in ${this.pluginName}: ${BYTECODE_MISMATCH_ERR_MSG}`);
+        return;
+      }
+
+      if (responseData.contracts === undefined || responseData.contracts === null) {
+        logger.error("There was an error during public verification. There are no returned contracts.");
         return;
       }
 
@@ -86,7 +128,7 @@ export class TenderlyService {
           addresses += `${cont.contractName}, `;
         }
 
-        console.log(`Error in ${this.pluginName}: ${NO_NEW_CONTRACTS_VERIFIED_ERR_MSG}`, addresses);
+        logger.error(`Error in ${this.pluginName}: ${NO_NEW_CONTRACTS_VERIFIED_ERR_MSG}`, addresses);
         return;
       }
 
@@ -101,8 +143,9 @@ export class TenderlyService {
       }
       console.groupEnd();
     } catch (err) {
-      logApiError(err);
-      console.log(`Error in ${this.pluginName}: ${API_VERIFICATION_REQUEST_ERR_MSG}`);
+      const logCompliantApiError = convertToLogCompliantApiError(err);
+      logger.error(logCompliantApiError);
+      logger.error(`Error in ${this.pluginName}: ${API_VERIFICATION_REQUEST_ERR_MSG}`);
     }
   }
 
@@ -111,8 +154,9 @@ export class TenderlyService {
     tenderlyProject: string,
     username: string
   ): Promise<void> {
+    logger.debug("Pushing contracts onto Tenderly.");
     if (!TenderlyApiService.isAuthenticated()) {
-      console.log(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
+      logger.error(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
       return;
     }
 
@@ -121,10 +165,16 @@ export class TenderlyService {
       const res = await tenderlyApi.post(`/api/v1/account/${username}/project/${tenderlyProject}/contracts`, {
         ...request,
       });
+      if (res.data === undefined || res.data === null) {
+        logger.error("There was an error while pushing contracts to Tenderly. Obtained response is invalid.");
+        return;
+      }
+      const logCompliantVerificationResponse = convertToLogCompliantVerificationResponse(res.data);
+      logger.trace("Verification response:", logCompliantVerificationResponse);
 
       const responseData: ContractResponse = res.data;
       if (responseData.bytecode_mismatch_errors !== null) {
-        console.log(`Error in ${this.pluginName}: ${BYTECODE_MISMATCH_ERR_MSG}`);
+        logger.error(`Error in ${this.pluginName}: ${BYTECODE_MISMATCH_ERR_MSG}`);
         return;
       }
 
@@ -134,7 +184,7 @@ export class TenderlyService {
           addresses += `${cont.contractName}, `;
         }
 
-        console.log(`Error in ${this.pluginName}: ${NO_NEW_CONTRACTS_VERIFIED_ERR_MSG}`, addresses);
+        logger.error(`Error in ${this.pluginName}: ${NO_NEW_CONTRACTS_VERIFIED_ERR_MSG}`, addresses);
         return;
       }
 
@@ -143,8 +193,9 @@ export class TenderlyService {
         `Successfully privately verified Smart Contracts for project ${tenderlyProject}. You can view your contracts at ${dashLink}`
       );
     } catch (err) {
-      logApiError(err);
-      console.log(`Error in ${this.pluginName}: ${API_VERIFICATION_REQUEST_ERR_MSG}`);
+      const logCompliantApiError = convertToLogCompliantApiError(err);
+      logger.error(logCompliantApiError);
+      logger.error(`Error in ${this.pluginName}: ${API_VERIFICATION_REQUEST_ERR_MSG}`);
     }
   }
 
@@ -154,8 +205,10 @@ export class TenderlyService {
     username: string,
     fork: string
   ): Promise<void> {
+    logger.info("Verifying contracts on fork.");
+
     if (!TenderlyApiService.isAuthenticated()) {
-      console.log(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
+      logger.error(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
       return;
     }
 
@@ -164,10 +217,15 @@ export class TenderlyService {
       const res = await tenderlyApi.post(`/account/${username}/project/${tenderlyProject}/fork/${fork}/verify`, {
         ...request,
       });
+      if (res.data === undefined || res.data === null) {
+        logger.error("There was an error while verifying contracts on fork. Obtained response is invalid.");
+      }
+      const logCompliantVerificationResponse = convertToLogCompliantVerificationResponse(res.data);
+      logger.trace("Verification response:", logCompliantVerificationResponse);
 
       const responseData: ContractResponse = res.data;
       if (responseData.bytecode_mismatch_errors !== null) {
-        console.log(BYTECODE_MISMATCH_ERR_MSG);
+        logger.error(BYTECODE_MISMATCH_ERR_MSG);
         return;
       }
 
@@ -177,7 +235,7 @@ export class TenderlyService {
           addresses += `${cont.contractName}, `;
         }
 
-        console.log(`Error in ${this.pluginName}: ${NO_NEW_CONTRACTS_VERIFIED_ERR_MSG}`, addresses);
+        logger.error(`Error in ${this.pluginName}: ${NO_NEW_CONTRACTS_VERIFIED_ERR_MSG}`, addresses);
         return;
       }
 
@@ -187,44 +245,62 @@ export class TenderlyService {
       }
       console.groupEnd();
     } catch (err) {
-      logApiError(err);
-      console.log(`Error in ${this.pluginName}: ${API_VERIFICATION_REQUEST_ERR_MSG}`);
+      const logCompliantApiError = convertToLogCompliantApiError(err);
+      logger.error(logCompliantApiError);
+      logger.error(`Error in ${this.pluginName}: ${API_VERIFICATION_REQUEST_ERR_MSG}`);
     }
   }
 
   public async getPrincipal(): Promise<Principal | null> {
+    logger.debug("Getting principal.");
+
     if (!TenderlyApiService.isAuthenticated()) {
-      console.log(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
+      logger.error(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
       return null;
     }
 
     const tenderlyApi = TenderlyApiService.configureInstance();
     try {
       const res = await tenderlyApi.get("/api/v1/user");
+      if (res.data === undefined || res.data === null) {
+        logger.error("There was an error while obtaining principal from Tenderly. Obtained response is invalid.");
+      }
+      logger.trace("Retrieved data:", { id: res.data.user.id });
+
       return {
         id: res.data.user.id,
         username: res.data.user.username,
       };
     } catch (err) {
-      logApiError(err);
-      console.log(`Error in ${this.pluginName}: ${PRINCIPAL_FETCH_FAILED_ERR_MSG}`);
+      const logCompliantApiError = convertToLogCompliantApiError(err);
+      logger.error(logCompliantApiError);
+      logger.error(`Error in ${this.pluginName}: ${PRINCIPAL_FETCH_FAILED_ERR_MSG}`);
     }
     return null;
   }
 
   public async getProjectSlugs(principalId: string): Promise<Project[]> {
+    logger.debug("Getting project slugs.");
+
     if (!TenderlyApiService.isAuthenticated()) {
-      console.log(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
+      logger.error(`Error in ${this.pluginName}: ${ACCESS_TOKEN_NOT_PROVIDED_ERR_MSG}`);
       return [];
     }
 
     const tenderlyApi = TenderlyApiService.configureInstance();
     try {
       const res = await tenderlyApi.get(`/api/v1/account/${principalId}/projects`);
+      if (res.data === undefined || res.data === null) {
+        logger.error("There was an error while obtaining project slug from Tenderly. Obtained response is invalid.");
+      }
+      const logCompliantProjects = convertToLogCompliantProjects(res.data.projects);
+      logger.trace("Obtained projects:", logCompliantProjects);
+
       return res.data.projects;
     } catch (err) {
-      logApiError(err);
-      console.log(`Error in ${this.pluginName}: ${PROJECTS_FETCH_FAILED_ERR_MSG}`);
+      const logCompliantApiError = convertToLogCompliantApiError(err);
+      logger.error(logCompliantApiError);
+      logger.error(`Error in ${this.pluginName}: ${PROJECTS_FETCH_FAILED_ERR_MSG}`);
     }
     return [];
   }